+++ src/method-name.ts
@@ -1,0 +1,6 @@
+/**
+ * Type to explicitly define a method name within a source code.
+ * 
+ * @type {MethodName}
+ */
+export type MethodName = string;