+++ src/serializer-handler.ts
@@ -1,0 +1,6 @@
+/**
+ * Serializer handler which is called by serializer during serialization and deserialization.
+ * 
+ * @type {SerializerHandler<TObject>}
+ */
+export type SerializerHandler<TObject> = (object: TObject) => void;