+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.2.0] - 2025-06-03
+
+### Added
+
+- Type option to specify the callback to invoke before serialization starts.
+- Type option to specify the callback to invoke when the deserialization is completed.
+
 ## [8.1.3] - 2025-06-02
 
 ### Changed